requirements.txt
streamlit
azure-data-tables
azure-storage-blob
azure-identity
pandas

